Move tests to same class.
